refactor(sdk): declare read-state and archive-coverage types in core/types
`core/types` is the SDK's leaf layer: the vocabulary every other layer names
its own contracts in. Six read-position types and three archive-coverage types
were declared under `stores/shared/`, so `core/types/chat.ts`, `room.ts` and
`sdk-events.ts` had to import from a state implementation to name them. That
made `core/types` depend on the stores that depend on it.

Move the declarations to `core/types/readState.ts` (new) and
`core/types/pagination.ts`. The `stores/shared` modules keep the constructors
and comparators and re-export the types, so no consumer changes. The public
type surface is unchanged.

This also removes the `readPointer` / `readState` import cycle, which existed
only because each module declared a type the other needed.

`client.ts` still derives `StoreBindings` from the concrete store state types;
`core/types/layering.test.ts` allows exactly that file and fails on any new
store import from `core/types`.
